Add files via upload
# Hw1
What do the numbers inside the parentheses after a function name do?
The numbers give an argument that the commond has to obey. The argument usually has the x and y axis to set up the initrial location. And then the numbers shows the weith and highth, even radious and so on.
What’s the difference between fill and stroke?
Fill is the interior of the shape, and the stroke is the outline of the shape.
The P5 web editor has a built-in console — but it turns out that Chrome also has a console to display errors. Do some googling to find out how you can open the Chrome console. Then, try to find a website that has errors — visit a website, open the console, and see if there are any red-highlighted items. Take a screenshot and upload it into your repository.

What is the keyboard shortcut to open the JavaScript console in Chrome on your system?
Ctrl Shift J
To open the developer console window on Chrome, use the keyboard shortcut Ctrl Shift J (on Windows) or Ctrl Option J (on Mac).